refactor: Reorganize folder structure and fix visual elements
changed folder structure to improve project organization and fixed visual elements, including: corrected alignment of the title, subtitle and logo in login form.
