--- assets/styles/global.css
@@ -1,3 +1,0 @@
-body{
-    margin: 0;
-}
